fix: :bug: fix bugs of spectator mode
